chore(code): reorganize code into multiple files
